qcacmn: Fix compile issue when not enabling ENH CFR
The extract_cfr_phase_param of the wmi_ops structure is
only defined when WLAN_ENH_CFR_ENABLE is enabled.
However, references to this member were not properly
guarded, leading to compilation errors when the macro
is disabled.

This change adds appropriate WLAN_ENH_CFR_ENABLE guards
to the references of extract_cfr_phase_param.

Change-Id: I036b7c19a8274dcca6ec34704d4bc17f37191ce4
CRs-Fixed: 4270989