qcacmn: Fix out of bounds read in extract_roam_scan_ap_stats_tlv
In API extract_roam_scan_ap_stats_tlv(), for loop is implemented
to extract AP info from a param_buf structure and store it in
a dst buffer starting from index value ap_idx.
The loop iterates num_cand times, where num_cand is the number
of candidate APs to be extracted. However, the issue arises
when the num_cand value exceeds the remaining number of APs
in the param_buf structure, starting from the ap_idx index.
This can cause the loop to access memory outside the bounds
of the param_buf structure.

To fix this, add a check before the for loop to ensure that
the num_cand value does not exceed the remaining number of
APs in the param_buf structure,starting from the
ap_idx index.

CRs-Fixed: 4218517
Change-Id: I46504dfd17da652fddd0bcea2f3f062420b3a9ff