qcacmn: Fix connection to low score candidate
Scenario:
1) AP is 2 + 6 MLO. Only the 6 GHz scan entry is present
in the DUT.
2) DUT attempts MLO connection to 6 GHz entry and auth fails.
3) Driver refreshes the candidate list during next connection
attempt. Now, the 2.4 GHz scan entry generated as part of the
ML-probe gets added as the next candidate to the candidate list.
4) DUT connects to 2.4 GHz candidate, even though 6 GHz SLO and
6 GHz 11ax candidates have higher score.

To fix this:
Choose the new entry as the next candidate only if it has
better score than the previously failed candidate. Or if
the BSSID matches with the self mac.
Otherwise, add the new entry based on the score.

Change-Id: I756a136366658b281f60b856373637ef94fd06c2
CRs-Fixed: 4229696