qcacmn: Prevent duplicate Tx descriptor processing
In the Tx completion path duplicate entries in the completion
ring can lead to NULL pointer dereference and system crash.This
occurs when the same descriptor is processed more than once.

This change adds a check to detect and skip duplicate entries
within the quota triggering an assertion for debugging.

CRs-Fixed: 4203198
Change-Id: I69e8d0de442ab14576a7651aad571c0400d3403a