qcacmn: Fix filter removal in opt_dp
It has been observed that the filter removal
operation is being interpreted as a filter
addition in the FW due to the 'valid' flag
not being explicitly set to 0.

This change is to set valid as false in
case of filter remove and set it back to
true if filter remove fails from FW end.

Change-Id: I680092642853affe7af572762092fcecd9cdbd25
CRs-Fixed: 4234708