qcacmn: Refine MSI configuration for MONITOR_STATUS
Currently as a part of INIT sequence, the interface flags (other_bss)
has an impact on whether MSI is configured for monitor status ring or
not. This dependency on interface flags can cause a scenario where no
MSI gets configured for monitor status ring, in the below scenario
1) Turn on wifi and create RxMon interface (other_bss=1)
2) Inject SSR (other_bss=1 is cached across SSR)
3) MSI configuration is skipped since other_bss=1
4) STA and RxMON interfaces are created after SSR.
5) Delete RxMON interface
6) Start LPC

Since MSI configuration does not depend on the use-case
to be executed after driver INIT, remove the check for
other_bss flag when deciding MSI configuration is required or not.

CRs-Fixed: 4214480
Change-Id: Ic03784ca033aa6ff1cc86973c3aa4ce40caeae81