qca-wifi-oss: Cancel CAC timer for punctured channel only when CAC timer is running
When STOP event is sent during DFS Punctured CAC Wait state, the CAC timer will be running and so it needs to be cancelled.
When STOP event is sent during DFS Punctured CAC Done state, the CAC timer will already be completed and stopped. So, its not required to cancel the timer at this stage.

Change-Id: I8c4e45d4320895214deca8b6953c046e21576342