qcacmn: Re-Enable runtime pm and low power on SSR/Driver unload
In the PCI case, runtime PM is disabled using force wake
and re-enabled based on a reference count.
However, in the IPCIE case, runtime PM is disabled using
a dedicated ID for opt_dp, and this ID is being
deregistered before runtime PM is re-enabled
using ref cnt, leading to deregistration failures in case
of SSR/Driver unload.

To fix this, runtime PM and low power modes are now
explicitly re-enabled during SSR and driver unload.
Additionally, the registration and deregistration of the
ID is moved to IPA UC attach/detach to ensure
runtime PM is properly re-enabled till the disable
count reaches zero.

Change-Id: Ic8dd1e1e743c8fbc7513531305d7e2a48a54e104
CRs-Fixed: 4241834