qca-wifi-oss: When a user attempts to puncture a channel that is already in the NOL, wait until the channel exits the NOL before initiating the CAC timer.
When a user attempts to puncture a channel that is already in the Non-Occupancy List (NOL), the CAC (Channel Availability Check) timer should not start immediately. Instead, wait until the channel exits the NOL before initiating the CAC timer.

Specifically, if a channel is in NOL due to radar detection and its puncture bitmap has been cleared, and the user tries to puncture the same channel again, the state machine should wait for the NOL expiry. Only after the channel exits NOL, the puncture state machine should go to CAC_WAIT state and start the CAC timer.

Change-Id: Ice0e908d6bcd7e871c0e3cbb475267effe034363
Signed-off-by: Ajithkumar Jayamoorthy <ajayamoo@qti.qualcomm.com>