qcacmn: Add notify NB in reg_remove_puncture()
Currently, lim_apply_puncture() is invoked when SAP starts or when the
channel changes, while lim_remove_puncture() is called when SAP stops.
When puncturing is applied, the is_static_punctured flag is set, and
the channel is disabled in the regulatory domain.

Although lim_apply_puncture() does not notify the wiphy channel list
about the disabled punctured channel, other events—such as Radar—do
trigger such notifications. Therefore, it is necessary to invoke the
notify NB callbacks in lim_remove_puncture() to ensure the removal
of puncture is properly communicated to the wiphy channel list.

Change-Id: I3b2a8c26d7f1e39598c6ce54b773ba8dc81fbc44
CRs-Fixed: 4253684