qcacmn: Avoid configuring filter pass mgmt filter in monitor mode
Currently for QCC2072, during monitor mode bringup, FP management
filters are set along with MO filters for the SW2RXDMA input ring.
This is overwriting the default filter pass configuration of routing
management frames to firmware. This is not a problem in standalone
monitor mode. But, in a STA+Monitor mode use case, this will result
in management packets routed to host driver instead of firmware,
creating issues wrt scan.

Avoid filter pass configuration for beacon and probe response frames
for SW2RXDMA input ring when monitor vdev is present. Since filter
pass configuration is not required for monitor mode, it is okay to
skip it in standalone monitor case as well.

CRs-Fixed: 4265067
Change-Id: If167d4d259e333eab3e6788ae5878a4f0eafca47