qcacmn: Fix the memory leak in duplicate frame detection
Currently, the driver allocates memory for the buffer upon
receiving a frame and releases it in case of failure. However,
when a duplicate frame is detected, the driver raises an error
but does not free the allocated memory. This leads to a memory
leak during stop modules.

So, fix this, ensure that buffer memory is properly freed in
all failure scenarios, including duplicate frame detection, to
prevent memory leaks.

Change-Id: Ica5ce033dbe7e3f96b243145af50e41b63f2452c
CRs-Fixed: 4279546