qcacmn: Remove RXDMA buffer filter for 2nd MAC
In mission mode, the second MAC is not used for RX sniffing.
Since the monitor status ring filter has already been removed
for MAC1, the RXDMA buffer filter for MAC1 is no longer necessary
and can be removed as well.

this is aimed to simplify the logic and avoid any issue due to MAC1
filter mismatch.

Change-Id: I582144334fb53b59ad502549045544227529fcc6
CRs-Fixed: 4327822