qcacmn: Prevent destruction of uninitialized REO command lock in fig DP
In the fig datapath, the REO command lock is currently not created.
however, the deinitialization process attempts to delete it, which
leads to a device crash.

To resolve this issue, move the lock deletion to the deinitialization,
counterpart of the lock creation api, ensuring that the lock is only
deleted if it is actually created.

Change-Id: Ic286f8fcbfaf7a4e1298b4b32881852fc1713276
CRs-Fixed: 4289152