qca-wifi:Skip CAC when secondary offsets are non-DFS channels during CSA to wider bandwidth
When AP operating on channel 52 with 80 MHz bandwidth triggers CSA to
channel 52 with 160 MHz, the host unnecessarily initiates CAC again—even
though the secondary offsets are non-DFS channels. This happens
because the new channel is not considered a subset of the old one.

To fix this, skip non-DFS channels during subset checks and
avoid CAC when secondary offsets are non-DFS.

Change-Id: I22f03a537641a0fe97cd23cc14c3c0588974339f
Signed-off-by: Divya R <divraj@qti.qualcomm.com>