qca-wifi: Fix for ACS and DCS choosing hw_blacklisted freq
When force ACS and DCS is given, ieee80211_find_dot11_channel
always choses the first channel in the curr_chan_list and
tries to come up on that channel. So, when the first channel is
blacklisted AP fails to come up on that channel.

Modify "reg_get_first_valid_freq" function return a frequency
that is not in the hw_blackisted channels list.

Change-Id: Ic08c40c70c183beb153b13772abea6bee447f5ce
Signed-off-by: Lokitha Thangavel <lthangav@qti.qualcomm.com>