qcacmn: Fix potential OOB read while extracting roam trigger TLV
The index used to access the roam trigger reason array is not
validated for bounds. This can lead to out-of-range access.

Add validation to ensure the index is less than the size of the roam
trigger reason array and return failure if the validation fails.

CRs-Fixed: 4309503
Change-Id: Ife9dea6e5c81cb5408b4f3b45ab68e652a5bab9f
(cherry picked from commit bbb11c160a0852ab449757254a1bac2ba3b671ca)