qcacmn: Resolve IPA Deinitialization State Transition
It is observed that filter addition triggered
during deinitialization is causing NULL
pointer access due to wait event being destroyed
while waiting a FW response for the addition.

To prevent this issue, the IPA state transition
to the deinitialization state has been moved
to the beginning of the deinit sequence.
This change ensures that filter additions are
blocked once the deinitialization stated, avoiding
unsafe operations.

CRs-Fixed: 4322032
Change-Id: Ib201cf83ddaf7bef4368a295b8dfe44a30d0c83d