qca-wifi-oss: Check explicitly for scan radio capability
The change 'I97a05777058ffc604bf2154f7f8497101ddb6ffe' modified the
behaviour of the API 'dfs_get_enable_york_dfs' only based on the INI
config alone and not based on the scan radio capability. This was done
to ensure that scan radio capability which is uninitialized is not
used incorrectly.

So, to support the changes made in the mentioned commit, add explicit
check for scan radio capability when radar is detected on a scan radio.

Change-Id: I651cb7c2c31fcb9d5071f1d1af1cefbb58c2c9e6