qcacmn: Add non-zero value check for TX deferred timestamp
The previous race condition fix
I07d8c8cb8dff109b900398418c4f83140fa4653f for in the TX path
prevents a system crash due to premature freeing of a descriptor
that hardware may still reference. The fix introduced deferred
timeout check but it does not check for default case when the
deferred timestamp is zero. The fix gets triggered every time
during start due to this.

A non-zero value check for the deferred timestamp helps target
the intended cases for the race condition fix.

Change-Id: I34e707ac295a62308b9b97b8b89cf541ba4bb48e
CRs-Fixed: 4331216