qcacmn: Avoid sending PMKID in WPA2 association requests
Currently, the association request frame for WPA2 connections
includes a cached or previous PMKID. This causes the AP to reject
the connection with an unspecified failure status code, leading to
connection issues.

To resolve this, skip sending the PMKID in WPA2 association
requests, especially when PMKSA is zero. This ensures compatibility
with APs that expect no PMKID in WPA2 scenarios and prevents
unnecessary connection failures.

Change-Id: Ia918709a6430b881cd9bfa05b3224d2b7fddd2b6
CRs-Fixed: 4329199