qcacmn: Add dp soc arch op to process nbuf list for DAL
Add DP arch op to process the nbuf list extracted from the RX
descriptors. This is required for the DAL RX functionality.

NBUFs extracted from the RX desc are enqueued into a list, since
the processing of the NBUFs require to invoke arch specific APIs,
it is required to have arch ops for the NBUF list processing.

Change-Id: I146c2ccbbb35be65ea21d3a781a5b9f08c69363a
CRs-Fixed: 4359429