qcacmn: Fix buffer overflow and enhance DC logging
In DC logging the character limit is not validated
before copying data into the log element buffer.
This is causing the buffer to overflow into the
next list element, corrupting the node and leading
to a page fault.

This change adds the following
1. Fix above page fault issue by validating payload size
before copying to ensure total size does not exceed
max allowed length.
2. Change max buffer size to MAX_PAYLOAD_SIZE + MAX_HEADER_SIZE
3. Introduce new macro to support rate limit in DC logging.
4. Log opt_dp pkt in both DC and Host log when encountered in
regular RX path.

Change-Id: I469f79c5b7e6ea704703507e131b058c20fad8e4
CRs-Fixed: 4350907