qcacmn: Change the Scan event handler context
Firmware uploads the beacons to the SRNG and sends the reap
event to host as part of the scan. The scan events like
foreign channel enter, scan complete are sent via the
WMI_SCAN_EVENTID.
WMI_SCAN_EVENTID is handled in tasklet context.
Reap event is handled in tasklet context, but after extracting the
beacon data from the SRNG, it posts a WMI_MGMT_RX_EVENTID which
is handled in the worker thread context.
Since the tasklet context gets prioritized first, few beacons
corresponding to the scan event are reaped after the scan
complete event is processed first.
In some order critical use cases like RRM, this results in omission
of the beacon to be reported to the AP, which was reaped after the
scan completion.

To fix this, handle the scan event also in the worker context
to maintain the order of processing.

CRs-Fixed: 4351923
Change-Id: I41c99495b3eaf528c67dfea0f51973877e7118f4