qcacmn: Handle direct refill ring setup
The feature capability setting based on HOST and FW
capability exchange can be delayed and occur after the
rx_mac_buf_rings are setup. Due to this, the direct refill
ring setup will be missed.

Handle the above mentioned scenario by doing the direct
refill ring setup when the feature is enabled after the
HOSt and FW capability exchange.

Change-Id: Ibc5f91f2e888fbf6da1b3de1aacb9e21af305268
CRs-Fixed: 4367568