qcacmn: Save the new link address during link switch abort
Race condition scenario between link switch and heart beat
timeout:
1) Link switch disconnect is in progress for a vdev.
2) FW sends heart beat timeout indication to the host driver.
3) The disconnect indication due to #2 is posted to scheduler.
4) Post LS disconnect completion, mac address of the standby link
is plumbed for the LS vdev.
3) Driver process the deauth indication posted at #3 from scheduler
and aborts the link switch SM.
4) set_mac response from the firmware gets terminated due to ABORT
state in the SM.

Firmware has successfully plumbed new mac address for the LS vdev,
but the new mac is not cached in the link info of ml_dev_ctx and
adapter of the driver. Therefore, if the next connection is attempted
without set_mac from the framework, then host and firmware goes
out-of-sync for this vdev. This leads to undesired effects.

To fix this, always update the ml_dev_ctx and adapter with the new
mac address whenever the set_mac response is successful from the
firmware.

Also, enhance the link switch SM logging.

Change-Id: Iae70c91011a4d5debffd689fd0e56e2d08d138f2
CRs-Fixed: 4368948