qcacmn: Set da_war_enabled for QCA6490
On QCA6490 with CONFIG_QCACLD_FEATURE_SON=y, it's found that
even after driver deletes AST entry for WDS peer, sa_is_valid
is still indicated as true on RX packet TLVs. Checking current
driver implementations, there is one similar fix that is added
for QCA8074. However the fix is implemented with the check of
ast_override_support. For QCA6490, it is set to true while it
is false for QCA8074. Simply setting ast_override_support to
false for QCA6490 is not an optimum choice, since it is referenced
in other code paths. Going through dp_soc structure, field
da_war_enabled is better, considering that it is referenced
in only one path in dp_rx_da_learn() without affect behaviors
of both QCA6490 and QCA8074.

With above, in this change, two changes are made.
1. Set soc->da_war_enabled to true for QCA6490 when
CONFIG_QCACLD_FEATURE_SON=y.

2. In dp_rx_wds_add_or_update_ast(), check soc->da_war_enabled
for the sa_is_valid issue which can be applicable to both
QCA6490 and QCA8074.

CRs-Fixed: 4377504
Change-Id: Ieed1572a182e322f39a7a3db6633816c0aa59b30