qcacmn: Assign more weightage to EHT non ML than HE candidate
If the connecting BSSID is on the avoid list, EHT non-ML and HE
candidates are assigned the same score, which could result in HE
candidates being preferred over EHT non-ML candidates.

To address this, assign greater weight to EHT non-ML candidates.
Additionally, consider the number of links and the phymode when
determining the better candidate while sorting.

Change-Id: If19dd39d0ef906ab915642a9728a1ac69f102c57
CRs-Fixed: 4378927