qcacmn: Add support to reset scan reject params
In instance of roam invoke failure or roam invoke abort the scan
reject params are not reset, this may result in HOST critical timeout
if threshold value is reached and needs to be reset in case of roam
abort event.
Adding a callback API to reset the scan reject params.

CRs-Fixed: 4389539
Change-Id: Ia7a001cd741fdb92de16906d145baa2d13c895ad