qcacmn: Fix rx reaping issue in dal rx handler
In the current logic, dal platform->rx() is invoked from the napi
handler during the first iteration always. But when napi execution
is delayed, there would not be any pending rx entries in the dal
module but there would be entries in the rx desc head in dp layer.

When platform->rx is invoked without checking the existing pending
descriptors to process, it would return zero pending entries in the
dal rings and the rx processing loop ends there as if no other entries
are pending to be processed. This stalls rx traffic.

Fix this by first processing entries present in the rx descriptor list
in the dp layer before fetching new entries from dal.

Change-Id: Iba2f89cdb6e6b6d835cc19c498509219da12dd11
CRs-Fixed: 4393895